+++ include/Config/HttpConfig.hpp
@@ -6,7 +6,7 @@
 /*   By: gyoon <gyoon@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/18 22:46:19 by gyoon             #+#    #+#             */
-/*   Updated: 2024/02/11 22:39:17 by gyoon            ###   ########.fr       */
+/*   Updated: 2024/02/15 15:07:33 by gyoon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 #define HTTPCONFIG_HPP
 
 #include "Config/AConfig.hpp"
-#include "Config/AHttpConfigModule.hpp"
 #include "Config/ConfigException.hpp"
+#include "Config/HttpConfigCore.hpp"
 #include "Config/ServerConfig.hpp"
 #include "File/ConfigFile.hpp"
 #include <string>
@@ -23,7 +23,7 @@
 
 namespace Hafserv
 {
-class HttpConfig : public AConfig, public AHttpConfigModule
+class HttpConfig : public AConfig, public HttpConfigCore
 {
 public:
 	HttpConfig();
