+++ src/Config/HttpConfigCore.cpp
@@ -6,7 +6,7 @@
 /*   By: gyoon <gyoon@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/19 16:26:26 by gyoon             #+#    #+#             */
-/*   Updated: 2024/02/14 20:55:24 by gyoon            ###   ########.fr       */
+/*   Updated: 2024/02/15 14:45:54 by gyoon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -96,6 +96,141 @@ void HttpConfigCore::addAllowMethod(const std::string &method)
 {
 	if (std::find(allowMethods.begin(), allowMethods.end(), method) == allowMethods.end())
 		allowMethods.push_back(method);
+}
+
+void HttpConfigCore::setHttpConfigCore(const ConfigFile::directives_t &directives)
+{
+	std::vector<std::string> params;
+	size_t numToken;
+	bool hasRoot = false;
+	ConfigFile::directives_t::const_iterator it = directives.begin();
+	for (; it != directives.end(); it++)
+	{
+		const std::string &key = (*it).first;
+		const std::string &value = (*it).second;
+
+		params = util::string::split(value, ' ');
+		numToken = params.size();
+		if (key == "client_body_buffer_size")
+		{
+			if (numToken != 1)
+				throw InvalidNumberArgumentError(key);
+			else if (!util::string::stoi(value).first)
+				throw InvalidArgumentError(key);
+
+			clientBodyBufferSize = util::string::stoi(value).second;
+		}
+		else if (key == "autoindex")
+		{
+			if (numToken != 1)
+				throw InvalidNumberArgumentError(key);
+
+			if (value == "on")
+				autoIndex = true;
+			else if (value == "off")
+				autoIndex = false;
+			else
+				throw InvalidArgumentError(key, value);
+		}
+		else if (key == "root")
+		{
+			if (hasRoot)
+				throw DuplicateDirectiveError(key);
+			else if (numToken != 1)
+				throw InvalidNumberArgumentError(key);
+
+			root = value;
+			hasRoot = true;
+		}
+		else if (key == "index")
+		{
+			std::vector<std::string> newIndexes;
+			for (size_t i = 0; i < params.size(); i++)
+				newIndexes.push_back(params[i]);
+
+			indexes = newIndexes;
+		}
+		else if (key == "client_header_timeout")
+		{
+			if (numToken != 1)
+				throw InvalidNumberArgumentError(key);
+			else if (!util::string::stoi(value).first)
+				throw InvalidArgumentError(key);
+
+			timeouts.clientHeader = util::string::stoi(value).second;
+		}
+		else if (key == "client_body_timeout")
+		{
+			if (numToken != 1)
+				throw InvalidNumberArgumentError(key);
+			else if (!util::string::stoi(value).first)
+				throw InvalidArgumentError(key);
+
+			timeouts.clientBody = util::string::stoi(value).second;
+		}
+		else if (key == "keepalive_timeout")
+		{
+			if (numToken != 1)
+				throw InvalidNumberArgumentError(key);
+			else if (!util::string::stoi(value).first)
+				throw InvalidArgumentError(key);
+
+			timeouts.keepAlive = util::string::stoi(value).second;
+		}
+		else if (key == "send_timeout")
+		{
+			if (numToken != 1)
+				throw InvalidNumberArgumentError(key);
+			else if (!util::string::stoi(value).first)
+				throw InvalidArgumentError(key);
+
+			timeouts.send = util::string::stoi(value).second;
+		}
+		else if (key == "error_page")
+		{
+			// what if error_page option duplicate?
+			if (params.size() == 1)
+				throw InvalidArgumentError(key, value);
+			for (size_t i = 0; i < params.size() - 1; i++)
+			{
+				if (util::string::stoi(params[i]).first && 300 <= util::string::stoi(params[i]).second &&
+					util::string::stoi(params[i]).second < 600)
+					errorPages.insert(std::make_pair(util::string::stoi(params[i]).second, params.back()));
+				else
+					; // error
+			}
+		}
+		else if (key == "allow_methods")
+		{
+			for (size_t i = 0; i < numToken; i++)
+			{
+				if (params[i] != "GET" && params[i] != "POST" && params[i] != "DELETE")
+					throw InvalidArgumentError(key, value);
+				addAllowMethod(params[i]);
+			}
+		}
+	}
+}
+
+void HttpConfigCore::setHttpConfigCore(const ConfigFile::subblocks_t &subBlocks)
+{
+	std::string subBlockName, type, extension;
+	for (size_t i = 0; i < subBlocks.size(); i++)
+	{
+		subBlockName = subBlocks.at(i).getBlockDirective();
+		if (subBlockName == "types")
+		{
+			ConfigFile::directives_t::const_iterator it = subBlocks.at(i).getDirectives().begin();
+			for (; it != subBlocks.at(i).getDirectives().end(); it++)
+			{
+				type = (*it).first;
+				extension = (*it).second;
+				std::vector<std::string> extensions = util::string::split(extension, ' ');
+				for (size_t j = 0; j < extensions.size(); j++)
+					addType(extensions.at(j), type);
+			}
+		}
+	}
 }
 
 std::ostream &operator<<(std::ostream &os, const HttpConfigCore &conf)
